Use form based login instead of HTTP basic auth
Form based login allows customization of the login form and adding a logo and explanation.
Session based login also allows for reliable ending of session. The previous logout implementation still had holes.

Implemented login as private action that gets called for any URI if the user is not authenticated, so the nice usability of basic auth should remain.